requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.1
scikit-learn==1.5.2
xgboost==2.1.1
joblib==1.4.2
pydantic==2.9.2
pydantic-settings==2.5.2
requests==2.32.3
redis==5.1.1
kafka-python==2.0.2
python-dotenv==1.0.1
plotly==5.24.1
pytest==8.3.3
httpx==0.27.2
